+++ src/main/java/me/flotsam/frettler/command/FrettlerCommand.java
@@ -52,6 +52,10 @@ public class FrettlerCommand {
   @Option(names = {"-i", "--intervals"}, description = "Show interval labels instead of notes")
   @Getter
   boolean intervals;
+
+  @Option(names = {"-l", "--list"}, description = "List chord definitions for the root (and interval pattern if provided)")
+  @Getter
+  boolean list;
   
   public enum View {
     HORIZONTAL, H(HORIZONTAL), VERTICAL, V(VERTICAL), DISPLAY, D(DISPLAY), FIND, F(FIND), CHORD, C(CHORD);